* Added skill INF2_NO_AUTOSHADOWSPELL for some skills that cannot be auto casted by Auto Shadow Spell as fix of bugreport:8582, http://rathena.org/board/tracker/issue-8582-shadow-chaserauto-shadow-spell-restricted-spells/ * Removed unused const.txt entries as fix of bugreport:8467, http://rathena.org/board/tracker/issue-8467-left-over-constants-from-mob-controller-system/

+++ src/map/pc.h
@@ -269,7 +269,7 @@ struct map_session_data {
 	uint16 skill_id_dance,skill_lv_dance;
 	short cook_mastery; // range: [0,1999] [Inkfish]
 	struct skill_cooldown_entry * scd[MAX_SKILLCOOLDOWN]; // Skill Cooldown
-	int cloneskill_idx, ///Stores index of copied skill by Intimidate/Plagiarism
+	short cloneskill_idx, ///Stores index of copied skill by Intimidate/Plagiarism
 		reproduceskill_idx; ///Stores index of copied skill by Reproduce
 	int menuskill_id, menuskill_val, menuskill_val2;
